Major refactor to get rid of templates and standardize everything to floats as much as possible. Moved a lot of header definitions to respective cpp files.

+++ Constants.h
@@ -10,9 +10,9 @@ class Constants final {
 
   static constexpr float FRAMERATE_60HZ_MS = 1000.0F / 60.0F;
 
-  static constexpr double PI = 3.14159265358979;
+  static constexpr float PI = 3.14159265358979f;
 
-  static constexpr double PI_OVER_180 = PI / 180.0;
+  static constexpr float PI_OVER_180 = PI / 180.0;
 
   static constexpr std::size_t MAX_VERTS_AFTER_CLIP = 2;
   static constexpr std::size_t MAX_INDICIES_AFTER_CLIP = 4;
